+++ build.gradle
@@ -4,7 +4,8 @@ plugins {
     id 'io.spring.dependency-management' version '1.0.15.RELEASE'
 }
 
-group = 'com.zerobase.cms.user'
+group = 'com.zerobase.cms'
+
 
 repositories {
     mavenCentral()
@@ -49,3 +50,6 @@ dependencyManagement {
 tasks.named('test') {
     useJUnitPlatform()
 }
+
+sourceCompatibility = '11'
+targetCompatibility = '11'
