+++ settings.gradle
@@ -1,1 +1,2 @@
-rootProject.name = 'user_api'
+rootProject.name = 'cms'
+include 'user_api', 'zerobase-domain'
