+++ src/main/java/com/zerobase/cms/main/CmsApplication.java
@@ -1,0 +1,22 @@
+package com.zerobase.cms.main;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+
+@EnableFeignClients
+@SpringBootApplication
+@RequiredArgsConstructor
+@EnableJpaAuditing
+@EnableJpaRepositories
+public class CmsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CmsApplication.class, args);
+    }
+
+}
