+++ src/main/java/com/zerobase/cms/main/client/MailgunClient.java
@@ -1,0 +1,18 @@
+package com.zerobase.cms.main.client;
+
+import com.zerobase.cms.main.client.mailgun.SendMailForm;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.cloud.openfeign.SpringQueryMap;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+
+@FeignClient(name="mailgun", url = "https://api.mailgun.net/v3/")
+@Qualifier("mailgun")
+public interface MailgunClient {
+
+    @PostMapping("sandboxb705fa342a4d4627bf5ee315a93aa1e8.mailgun.org/messages")
+    ResponseEntity<String> sendEmail(@SpringQueryMap SendMailForm form);
+
+
+}
