+++ src/main/java/com/zerobase/cms/main/client/mailgun/SendMailForm.java
@@ -1,0 +1,20 @@
+package com.zerobase.cms.main.client.mailgun;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Builder
+@Data
+public class SendMailForm {
+    private String from;
+    private String to;
+    private String subject;
+    private String text;
+
+}
