--- src/main/java/com/zerobase/cms/user/UserApiApplication.java
@@ -1,22 +1,0 @@
-package com.zerobase.cms.user;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-
-@EnableFeignClients
-@SpringBootApplication
-@RequiredArgsConstructor
-@EnableJpaAuditing
-@EnableJpaRepositories
-public class UserApiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(UserApiApplication.class, args);
-    }
-
-}
