--- src/main/java/com/zerobase/cms/user/client/mailgun/SendMailForm.java
@@ -1,20 +1,0 @@
-package com.zerobase.cms.user.client.mailgun;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Builder
-@Data
-public class SendMailForm {
-    private String from;
-    private String to;
-    private String subject;
-    private String text;
-
-}
