--- src/main/java/com/zerobase/cms/user/service/SignUpCustomerService.java
@@ -1,20 +1,0 @@
-package com.zerobase.cms.user.service;
-
-import com.zerobase.cms.user.domain.SignUpForm;
-import com.zerobase.cms.user.domain.model.Customer;
-import com.zerobase.cms.user.domain.repository.CustomerRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class SignUpCustomerService {
-
-    private final CustomerRepository customerRepository;
-
-    public Customer signUp(SignUpForm form){
-
-        return customerRepository.save(Customer.from(form));
-    }
-}
-
