+++ src/test/java/com/zerobase/cms/domain/CmsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.zerobase.cms.domain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CmsApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
