--- src/test/java/com/zerobase/cms/user/UserApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.zerobase.cms.user;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
