+++ zerobase-domain/src/main/java/com/zerobase/cms/domain/domain/common/UserType.java
@@ -1,0 +1,5 @@
+package com.zerobase.cms.domain.domain.common;
+
+public enum UserType {
+    CUSTOMER, SELLER;
+}
