+++ zerobase-domain/src/test/java/com/zerobase/cms/domain/util/Aes256UtilTest.java
@@ -1,0 +1,15 @@
+package com.zerobase.cms.domain.util;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
+
+class Aes256UtilTest {
+
+    @Test
+    void encrypt() {
+        String encrypt = Aes256Util.encrypt("Hello world");
+        assertEquals(Aes256Util.decrypt(encrypt), "Hello world");
+    }
+
+}
